pulling data from the open weather api

+++ app/src/main/java/ser421/edu/lab_6_native/WeatherReport.java
@@ -2,8 +2,8 @@ package ser421.edu.lab_6_native;
 
 public class WeatherReport {
     protected String location;
-    protected int temperature;
-    protected int humidity;
-    protected int windSpeed;
-    protected int cloudCover;
+    protected double temperature;
+    protected double humidity;
+    protected double windSpeed;
+    protected double cloudCover;
 }
